Fix fractional supercell orders truncated to int in parallel batch path
The parallel BatchSimulationRunner path serializes case results through
JSON records; deserialization cast the orders array back with dtype=int,
truncating fractional supercell orders (e.g. -1.2, -1.4, -1.6, -1.8 all
collapsed onto -1). The all-orders CSV then contained several near-zero
rows labeled order -1 per energy alongside the real one, which made the
comparison plot zigzag to zero between energy points for supercell runs.

Deserialize orders as float, apply the integer-preserving CSV order
formatting to the theta-search sweep writer too, and align the empty
placeholder arrays. Adds a round-trip regression test.

+++ src/grax/simulation/core.py
@@ -610,7 +610,7 @@ class RCWASimulation:
             )
             for energy in energies
         ]
-        orders = single_results[0].orders if single_results else np.asarray([], dtype=int)
+        orders = single_results[0].orders if single_results else np.asarray([], dtype=float)
         for result in single_results[1:]:
             if not np.array_equal(orders, result.orders):
                 raise ValueError("Diffraction orders changed between energy points.")
